MC estimation ala vineppo
